Qt: some work on upgrading to version 5.6 (#908) [ci skip].
The move to QTEST_APPLESS_MAIN() for our tests seemed sensible, but... this didn't account for the fact that we have some XML-based tests (COMBINESupport::tests and Core::mathmltests) and that this requires an application. So, rather than QTEST_APPLESS_MAIN(), we should use QTEST_GUILESS_MAIN() for at least those two tests, but in the end we might as well use QTEST_GUILESS_MAIN() for all of them (just in case for the future).

Note #1: we don't want to use QTEST_MAIN() since we want to be able to run our tests on Travis CI, which means that we cannot have a GUI.

Note #2: interestingly enough, the issue with our XML-based tests only occurs on Windows. In other words, to use QTEST_APPLESS_MAIN() on Linux and OS X is fine in the sense that the tests pass while they crash on Windows.