Qt: some work on upgrading to version 5.6 (#908) [ci skip].
All that time, I thought the issue with not being able to run OpenCOR
in my Windows VM was because of VirtualBox. In the end, it simply
proved to be a problem with the WebEngineWidgets module not being
linked in the OpenCOR executable!

Strangely enough, that wasn’t a problem on a ‘real’ machine, just on a
VM...!?